backend/requirements.txt
fastapi>=0.108.0
sqlmodel>=0.0.16
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0